Refactor tests a bit
